Init Commit
Used JavaScript to create a simple game. You can choose to fight or
run. If you run, an image switches out. If you fight, dice roll, and if
roll is greater than 9 you win. Wrote an if statement to prompt a win
or loss, but it messed up the display of the dice onscreen. Need to
come back style with CSS.